HWI/requirements.txt
gpiocrust
jsonrpclib-pelix